finished basic core

+++ modules-template/core/src/main/java/org/example/core/service/StudentService.java
@@ -1,12 +1,12 @@
 package org.example.core.service;
 
 import org.example.core.model.Student;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
-
 public interface StudentService {
   List<Student> findAll();
   boolean updateStudent(Integer id, String newName);
   boolean saveStudent(Integer id, String name);
-  boolean deleteStudent(Long id);
+  boolean deleteStudent(Integer id);
 }
